+++ include/workspace/snippetmgr.h
