+++ src/builtincmd.cpp
@@ -1,20 +1,57 @@
-#include "cmd/builtincmd.h"
+#include "builtincmd.h"
 
 #include <memory>
-#include "term.h"
 
-#include "cmd/commandmgr.h"
+#include "term.h"
 #include "console.h"
-#include "sapp/downloader.h"
-#include "utils/credit.h"
-#include "utils/utils.h"
+#include "downloader.h"
+#include "credit.h"
+#include "utils.h"
 #include "version.h"
 #include "snippetmgr.h"
+#include "sapp.h"
+
+using std::make_shared;
+
+void loadCommands() {
+    clearCommands();
+
+    commands.push_back(make_shared<BuiltinCommand>(BuiltinCommand(
+            "shfl", "Shuffle Command", {
+                    CommandOption("reload", BOOL_T, {"rl"}),
+                    CommandOption("update", BOOL_T, {"upgrade"}),
+                    CommandOption("credits", BOOL_T, {"credit"}),
+            }, shflCmd
+    )));
+    commands.push_back(make_shared<BuiltinCommand>(BuiltinCommand(
+            "appmgr", "App Manager", {
+                    CommandOption("add", TEXT_T, {"ad", "a"}),
+                    CommandOption("remove", TEXT_T, {"rm", "r"}),
+            }, appMgrCmd
+    )));
+    commands.push_back(make_shared<BuiltinCommand>(BuiltinCommand(
+            "help", "Show help", {
+                    CommandOption("command", TEXT_T, {"cmd", "help"})
+            }, helpCmd
+    )));
+    commands.push_back(make_shared<BuiltinCommand>(BuiltinCommand(
+            "snf", "Manage Snippets", {
+                    CommandOption("create", TEXT_T, {"mk", "c", "new"}),
+                    CommandOption("value", TEXT_T, {"v"}),
+            }, snippetCmd
+    )));
+
+    for (const CommandData &command: getRegisteredCommands()) {
+        loadApp(command);
+    }
+}
 
 void shflCmd(Workspace &ws, map<string, string> &optionValues) {
     if (optionValues["reload"] == "true") {
         info("Reloading command...");
+
         loadCommands();
+
         success("Reloaded all commands!");
     } else if (optionValues["apps"] == "true") {
         if (optionValues.count("add")) {
