+++ src/cmd/cmdparser.cpp
@@ -1,4 +1,4 @@
-#include "cmd/cmdparser.h"
+#include "cmdparser.h"
 
 #include <algorithm>
