+++ src/main.cpp
@@ -3,24 +3,24 @@
 #include <csignal>
 
 #include "console.h"
-#include "cmd/commandmgr.h"
-#include "utils/utils.h"
-#include "utils/crashreport.h"
+#include "utils.h"
+#include "crashreport.h"
 #include "version.h"
 #include "snippetmgr.h"
+#include "builtincmd.h"
 
 #ifdef _WIN32
 
 LONG WINAPI ExceptionHandler(EXCEPTION_POINTERS *exceptionPointers) {
-  term << newLine;
-  error("Sorry. Something went wrong with Shuffle. Go to the URL below and report the problem.");
-  error("https://github.com/shflterm/shuffle/issues/new?template=crash-report.yaml");
-  term << newLine;
-  CrashReport report = CrashReport()
-      .setStackTrace(genStackTrace(exceptionPointers->ContextRecord));
-  error(report.make());
-  report.save();
-  exit(EXCEPTION_CONTINUE_SEARCH);
+    term << newLine;
+    error("Sorry. Something went wrong with Shuffle. Go to the URL below and report the problem.");
+    error("https://github.com/shflterm/shuffle/issues/new?template=crash-report.yaml");
+    term << newLine;
+    CrashReport report = CrashReport()
+            .setStackTrace(genStackTrace(exceptionPointers->ContextRecord));
+    error(report.make());
+    report.save();
+    exit(EXCEPTION_CONTINUE_SEARCH);
 }
 
 #elif defined(__linux__) || defined(__APPLE__)
