+++ src/utils/CMakeLists.txt
@@ -1,24 +1,17 @@
 cmake_minimum_required(VERSION 3.18)
-project(shuffle)
+project(utils)
 
-set(APP_SOURCES main.cpp builtincmd.cpp)
+set(UTILS_SRC utils.cpp lua/luaapi.cpp credit.cpp console.cpp i18n.cpp)
 
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 
-add_executable(${PROJECT_NAME} ${APP_SOURCES})
+add_library(${PROJECT_NAME} ${UTILS_SRC})
 
-add_subdirectory(cmd)
-add_subdirectory(sapp)
-add_subdirectory(suggestion)
-add_subdirectory(workspace)
-add_subdirectory(utils)
-
-add_dependencies(${PROJECT_NAME} commandmgr sappmgr suggestion workspace utils)
-target_link_libraries(${PROJECT_NAME} PRIVATE commandmgr sappmgr suggestion workspace utils)
+add_dependencies(${PROJECT_NAME} term_cpp)
 
 target_compile_definitions(${PROJECT_NAME} PRIVATE _HAS_STD_BYTE=0)
 
-target_link_libraries(${PROJECT_NAME} PRIVATE term_cpp)
+target_link_libraries(${PROJECT_NAME} PUBLIC term_cpp)
 target_include_directories(${PROJECT_NAME} PRIVATE ../termcpp/include)
 
 find_package(jsoncpp CONFIG REQUIRED)
@@ -36,4 +29,4 @@ target_link_libraries(${PROJECT_NAME} PRIVATE kubazip::kubazip)
 
 target_link_directories(${PROJECT_NAME} PRIVATE ../build/lib)
 
-target_include_directories(${PROJECT_NAME} PRIVATE ../include ../include/cmd ../include/sapp ../include/suggestion ../include/workspace ../include/utils)
+target_include_directories(${PROJECT_NAME} PRIVATE ../../include ../../include/cmd ../../include/sapp ../../include/suggestion ../../include/workspace ../../include/utils)
