+++ src/workspace/snippetmgr.cpp
@@ -2,12 +2,12 @@
 // Created by 08sih on 6/25/2023.
 //
 
-#include "snippetmgr.h"
+#include "workspace/snippetmgr.h"
 
 #include <utility>
-#include <json/json.h>
+#include "json/json.h"
 
-#include "utils/utils.h"
+#include "utils.h"
 
 using std::vector, std::make_shared;
